cpp: bioformats: Move PlaneRegion out from tiff namespace
This isn't TIFF-specific and is useful in other readers and
writers.
