+++ cpp/lib/ome/bioformats/Types.h
@@ -71,4 +71,3 @@ namespace ome
  * mode:C++
  * End:
  */
-
