+++ cpp/lib/ome/bioformats/tiff/IFD.cpp
@@ -42,6 +42,7 @@
 
 #include <ome/internal/config.h>
 
+#include <ome/bioformats/PlaneRegion.h>
 #include <ome/bioformats/tiff/IFD.h>
 #include <ome/bioformats/tiff/Tags.h>
 #include <ome/bioformats/tiff/Field.h>
@@ -65,6 +66,8 @@ namespace
   using ::ome::bioformats::dimension_size_type;
   using ::ome::bioformats::PixelBuffer;
   using ::ome::bioformats::PixelProperties;
+  using ::ome::bioformats::PlaneRegion;
+  using ::ome::bioformats::TileBuffer;
 
   // VariantPixelBuffer tile transfer
   // ────────────────────────────────
