+++ cpp/test/ome-bioformats/tiff.cpp
@@ -65,11 +65,11 @@ using ome::bioformats::tiff::directory_index_type;
 using ome::bioformats::tiff::TileInfo;
 using ome::bioformats::tiff::TIFF;
 using ome::bioformats::tiff::IFD;
-using ome::bioformats::tiff::PlaneRegion;
 using ome::bioformats::dimension_size_type;
 using ome::bioformats::VariantPixelBuffer;
 using ome::bioformats::PixelBuffer;
 using ome::bioformats::PixelProperties;
+using ome::bioformats::PlaneRegion;
 typedef ome::xml::model::enums::PixelType PT;
 
 using namespace boost::filesystem;
@@ -297,205 +297,6 @@ namespace
 }
 
 std::vector<TileTestParameters> tile_params(find_tile_tests());
-
-TEST(TIFFTest, RegionIntersection1)
-{
-  PlaneRegion r1(0, 0, 50, 100);
-  PlaneRegion r2(25, 30, 100, 50);
-  PlaneRegion r3 = r1 & r2;
-  EXPECT_EQ(25U, r3.x);
-  EXPECT_EQ(30U, r3.y);
-  EXPECT_EQ(25U, r3.w);
-  EXPECT_EQ(50U, r3.h);
-  PlaneRegion r4 = r2 & r1;
-  EXPECT_EQ(25U, r4.x);
-  EXPECT_EQ(30U, r4.y);
-  EXPECT_EQ(25U, r4.w);
-  EXPECT_EQ(50U, r4.h);
-}
-
-TEST(TIFFTest, RegionIntersection2)
-{
-  PlaneRegion r1(0, 0, 100, 100);
-  PlaneRegion r2(25, 25, 25, 25);
-  PlaneRegion r3 = r1 & r2;
-  EXPECT_EQ(25U, r3.x);
-  EXPECT_EQ(25U, r3.y);
-  EXPECT_EQ(25U, r3.w);
-  EXPECT_EQ(25U, r3.h);
-  PlaneRegion r4 = r2 & r1;
-  EXPECT_EQ(25U, r4.x);
-  EXPECT_EQ(25U, r4.y);
-  EXPECT_EQ(25U, r4.w);
-  EXPECT_EQ(25U, r4.h);
-}
-
-TEST(TIFFTest, RegionIntersection3)
-{
-  PlaneRegion r1(40, 190, 29, 18);
-  PlaneRegion r2(40, 190, 29, 18);
-  PlaneRegion r3 = r1 & r2;
-  EXPECT_EQ(40U, r3.x);
-  EXPECT_EQ(190U, r3.y);
-  EXPECT_EQ(29U, r3.w);
-  EXPECT_EQ(18U, r3.h);
-  PlaneRegion r4 = r2 & r1;
-  EXPECT_EQ(40U, r4.x);
-  EXPECT_EQ(190U, r4.y);
-  EXPECT_EQ(29U, r4.w);
-  EXPECT_EQ(18U, r4.h);
-}
-
-TEST(TIFFTest, RegionIntersection4)
-{
-  PlaneRegion r1(20, 30, 80, 50);
-  PlaneRegion r2(200, 25, 60, 20);
-  PlaneRegion r3 = r1 & r2;
-  EXPECT_EQ(0U, r3.x);
-  EXPECT_EQ(0U, r3.y);
-  EXPECT_EQ(0U, r3.w);
-  EXPECT_EQ(0U, r3.h);
-  PlaneRegion r4 = r2 & r1;
-  EXPECT_EQ(0U, r4.x);
-  EXPECT_EQ(0U, r4.y);
-  EXPECT_EQ(0U, r4.w);
-  EXPECT_EQ(0U, r4.h);
-}
-
-TEST(TIFFTest, RegionIntersection5)
-{
-  PlaneRegion r1(20, 400, 80, 50);
-  PlaneRegion r2(30, 25, 60, 45);
-  PlaneRegion r3 = r1 & r2;
-  EXPECT_EQ(0U, r3.x);
-  EXPECT_EQ(0U, r3.y);
-  EXPECT_EQ(0U, r3.w);
-  EXPECT_EQ(0U, r3.h);
-  PlaneRegion r4 = r2 & r1;
-  EXPECT_EQ(0U, r4.x);
-  EXPECT_EQ(0U, r4.y);
-  EXPECT_EQ(0U, r4.w);
-  EXPECT_EQ(0U, r4.h);
-}
-
-TEST(TIFFTest, RegionUnion1)
-{
-  PlaneRegion r1(0, 0, 16, 16);
-  PlaneRegion r2(16, 0, 16, 16);
-  PlaneRegion r3 = r1 | r2;
-  EXPECT_EQ(0U, r3.x);
-  EXPECT_EQ(0U, r3.y);
-  EXPECT_EQ(32U, r3.w);
-  EXPECT_EQ(16U, r3.h);
-  PlaneRegion r4 = r2 | r1;
-  EXPECT_EQ(0U, r4.x);
-  EXPECT_EQ(0U, r4.y);
-  EXPECT_EQ(32U, r4.w);
-  EXPECT_EQ(16U, r4.h);
-}
-
-TEST(TIFFTest, RegionUnion2)
-{
-  PlaneRegion r1(0, 0, 16, 16);
-  PlaneRegion r2(0, 16, 16, 16);
-  PlaneRegion r3 = r1 | r2;
-  EXPECT_EQ(0U, r3.x);
-  EXPECT_EQ(0U, r3.y);
-  EXPECT_EQ(16U, r3.w);
-  EXPECT_EQ(32U, r3.h);
-  PlaneRegion r4 = r2 | r1;
-  EXPECT_EQ(0U, r4.x);
-  EXPECT_EQ(0U, r4.y);
-  EXPECT_EQ(16U, r4.w);
-  EXPECT_EQ(32U, r4.h);
-}
-
-TEST(TIFFTest, RegionUnion3)
-{
-  PlaneRegion r1(43, 23, 12, 15);
-  PlaneRegion r2(55, 23, 44, 15);
-  PlaneRegion r3 = r1 | r2;
-  EXPECT_EQ(43U, r3.x);
-  EXPECT_EQ(23U, r3.y);
-  EXPECT_EQ(56U, r3.w);
-  EXPECT_EQ(15U, r3.h);
-  PlaneRegion r4 = r2 | r1;
-  EXPECT_EQ(43U, r4.x);
-  EXPECT_EQ(23U, r4.y);
-  EXPECT_EQ(56U, r4.w);
-  EXPECT_EQ(15U, r4.h);
-}
-
-TEST(TIFFTest, RegionUnion4)
-{
-  PlaneRegion r1(22, 19, 27, 80);
-  PlaneRegion r2(22, 99, 27, 11);
-  PlaneRegion r3 = r1 | r2;
-  EXPECT_EQ(22U, r3.x);
-  EXPECT_EQ(19U, r3.y);
-  EXPECT_EQ(27U, r3.w);
-  EXPECT_EQ(91U, r3.h);
-  PlaneRegion r4 = r2 | r1;
-  EXPECT_EQ(22U, r4.x);
-  EXPECT_EQ(19U, r4.y);
-  EXPECT_EQ(27U, r4.w);
-  EXPECT_EQ(91U, r4.h);
-}
-
-TEST(TIFFTest, RegionUnion5)
-{
-  // No overlap or common edge
-  PlaneRegion r1(43, 23, 12, 15);
-  PlaneRegion r2(95, 83, 43, 15);
-  PlaneRegion r3 = r1 | r2;
-  EXPECT_EQ(0U, r3.x);
-  EXPECT_EQ(0U, r3.y);
-  EXPECT_EQ(0U, r3.w);
-  EXPECT_EQ(0U, r3.h);
-  PlaneRegion r4 = r2 | r1;
-  EXPECT_EQ(0U, r4.x);
-  EXPECT_EQ(0U, r4.y);
-  EXPECT_EQ(0U, r4.w);
-  EXPECT_EQ(0U, r4.h);
-}
-
-TEST(TIFFTest, RegionUnion6)
-{
-  // Overlap
-  PlaneRegion r1(43, 23, 12, 15);
-  PlaneRegion r2(50, 28, 12, 15);
-  PlaneRegion r3 = r1 | r2;
-  EXPECT_EQ(0U, r3.x);
-  EXPECT_EQ(0U, r3.y);
-  EXPECT_EQ(0U, r3.w);
-  EXPECT_EQ(0U, r3.h);
-  PlaneRegion r4 = r2 | r1;
-  EXPECT_EQ(0U, r4.x);
-  EXPECT_EQ(0U, r4.y);
-  EXPECT_EQ(0U, r4.w);
-  EXPECT_EQ(0U, r4.h);
-}
-
-TEST(TIFFTest, RegionValid)
-{
-  // Overlap
-  PlaneRegion r1;
-  ASSERT_TRUE(!r1.valid());
-  ASSERT_FALSE(r1.valid());
-
-  PlaneRegion r2(0, 0, 5, 2);
-  ASSERT_TRUE(r2.valid());
-  ASSERT_FALSE(!r2.valid());
-}
-
-TEST(TIFFTest, RegionArea)
-{
-  PlaneRegion r1;
-  ASSERT_EQ(0U, r1.area());
-
-  PlaneRegion r2(0, 0, 4, 2);
-  ASSERT_EQ(4U*2U, r2.area());
-}
 
 TEST(TIFFTest, Construct)
 {
